hubstaff/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1
django-cors-headers==4.4.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-extra-fields==3.7.0
EasyProcess==1.1
entrypoint2==1.1
evdev==1.7.1
filetype==1.2.0
jeepney==0.8.0
kombu==5.4.1
MouseInfo==0.1.3
mss==9.0.2
pillow==10.4.0
prompt_toolkit==3.0.47
psycopg==3.2.2
psycopg2-binary==2.9.9
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyJWT==2.9.0
PyMsgBox==1.0.9
pynput==1.7.7
pyperclip==1.9.0
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-xlib==0.33
python3-xlib==0.15
pytweening==1.2.0
redis==5.0.8
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
